Moved Contact change to be a separate component
Added back call to TaskList as it was lost in a previous commit from an outdated branch
