+++ src/views/ResLife/components/RAView/index.jsx
@@ -28,7 +28,8 @@ import {
   submitCheckIn,
   getRACurrentHalls,
 } from 'services/residentLife/RA_Checkin';
-import { preferredContact, PrefContactMethod } from 'services/residentLife/ResidentStaff';
+import ContactMethod from './components/ContactMethod';
+import TaskList from './components/TaskList';
 import { useUser } from 'hooks';
 import HousingBanner from '../ResidentView/components/HousingWelcome/Banner';
 import OnDutyMobile from '../RDView/components/OnDutyMobileView';
@@ -94,22 +95,6 @@ const RAView = () => {
       }
     };
     fetchData();
-  }, [profile?.ID]);
-
-  //Auto set radio button in pref contact card to current pref method
-  useEffect(() => {
-    const fetchPreferredContact = async () => {
-      if (profile?.ID) {
-        try {
-          const contactPreference = await PrefContactMethod(profile.ID);
-          setSelectedContact(contactPreference?.PreferredContactMethod || '');
-        } catch (error) {
-          console.error('Error fetching preferred contact method:', error);
-        }
-      }
-    };
-
-    fetchPreferredContact();
   }, [profile?.ID]);
 
   const hallCodeToNameMap = {
@@ -217,25 +202,6 @@ const RAView = () => {
 
       return updatedState;
     });
-  };
-
-  const handleContactChange = (event) => {
-    setSelectedContact(event.target.value);
-  };
-
-  const handleContactSubmit = async () => {
-    if (!selectedContact) {
-      alert('Please select a contact method before submitting.');
-      return;
-    }
-
-    try {
-      await preferredContact(profile.ID, selectedContact);
-      alert('Preferred contact method successfully updated.');
-    } catch (error) {
-      console.error('Error updating preferred contact method:', error);
-      alert('Failed to update contact method. Please try again.');
-    }
   };
 
   const checkInButton = () => (
@@ -372,38 +338,6 @@ const RAView = () => {
     </Grid>
   );
 
-  const contactMethod = () => (
-    <Card elevation={3} sx={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
-      <CardHeader title={'Preferred Contact Method'} className="gc360_header" />
-      <CardContent sx={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
-        <FormControl component="fieldset">
-          <Typography>Select Contact Method</Typography>
-          <RadioGroup
-            aria-label="preferred-contact"
-            name="preferred-contact"
-            value={selectedContact}
-            onChange={handleContactChange}
-          >
-            <FormControlLabel value="teams" control={<Radio />} label="Teams" />
-            <FormControlLabel value="phone" control={<Radio />} label="Phone" />
-          </RadioGroup>
-          <Button
-            variant="contained"
-            color="primary"
-            onClick={handleContactSubmit}
-            disabled={!selectedContact}
-            sx={{ mt: 2 }}
-          >
-            Submit
-          </Button>
-        </FormControl>
-        <Typography sx={{ mt: 2, color: 'text.secondary', fontStyle: 'italic' }}>
-          *This is your preferred method to be contacted by your hall's residents.
-        </Typography>
-      </CardContent>
-    </Card>
-  );
-
   const OnCallTable = () => {
     return (
       <Grid item xs={12} md={20} padding={1}>
@@ -434,11 +368,16 @@ const RAView = () => {
           <Grid item xs={12} md={4}>
             <OnCallTable />
           </Grid>
-          {contactMethod()}
+          <ContactMethod />
           <Grid item xs={12} md={4}>
             <MyHall />
           </Grid>
           {checkInButton()}
+          {isCheckedIn ? (
+            <Grid item xs={12} md={4}>
+              <TaskList />
+            </Grid>
+          ) : null}
         </>
       )}
       {isMobile && (
@@ -450,7 +389,10 @@ const RAView = () => {
           <Grid item xs={12}>
             <MyHall />
           </Grid>
-          {contactMethod()}
+          <ContactMethod />
+          <Grid item xs={12}>
+            {isCheckedIn ? <TaskList /> : null}
+          </Grid>
           <Grid item xs={12}>
             <OnCallTable />
           </Grid>
